+++ changelog.md
@@ -3,8 +3,8 @@ All commits are documented in this file.
 
 ## [06/10/2020]
 ### Added 
- - set up node JS with express and postgress packages
- - basic starter environment
+ - postgres DB and express environment
+ - basic tables for api
  - changelog
 
  [06/10/2020]: https://github.com/Gbillington1/Events-API/compare/d3b101d..HEAD